--- src/assets/template/connection.html
@@ -1,150 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
-        <title>Document</title>
-
-        <style type="text/css">
-        </style>
-
-    </head>
-    <body>
-
-        <template data-template="connection-item">
-            <div>
-                <h3>Label</h3>
-                <dl>
-                    <dt>Host</dt>
-                    <dd data-section="host"></dd>
-                    <dt>Path</dt>
-                    <dd data-section="path"></dd>
-                    <dt>Port</dt>
-                    <dd data-section="port"></dd>
-                    <dt>Secure</dt>
-                    <dd data-section="secure"></dd>
-                </dl>
-                <button type="button" data-action="delete">Delete</button>
-            </div>
-        </template>
-
-        <h3>Connections</h3>
-
-        <div class="connections" data-section="list">
-        </div>
-
-        <hr >
-
-        <h3>Add new Connection</h3>
-
-        <!-- add form -->
-        <form data-section="add-form">
-            <div class="control" data-control="host">
-                <label for="host">Host</label><input type="text" id="host" />
-            </div>
-            <div class="control" data-control="path">
-                <label for="path">Path</label><input type="text" id="path" />
-            </div>
-            <div class="control" data-control="port">
-                <label for="port">Port</label><input type="number" id="port" />
-            </div>
-            <div class="control" data-control="secure">
-                <label for="path">Secure Connection</label>
-                <input type="checkbox" checked="checked">
-            </div>
-            <button type="submit">Save</button>
-        </form>
-
-        <script type="text/javascript">
-            const vscode       = acquireVsCodeApi();
-            const list         = document.querySelector('div[data-section="list"]');
-
-            /** form */
-            const form         = document.querySelector('[data-section="add-form"');
-            const controls     = Array.from(document.querySelectorAll('form [data-control]').values());
-
-            /** list */
-            const itemTemplate = document.querySelector('[data-template="connection-item"]');
-            const template = new Map();
-            template.set('host', itemTemplate.content.querySelector('[data-section="host"]'));
-            template.set('path', itemTemplate.content.querySelector('[data-section="path"]'));
-            template.set('port', itemTemplate.content.querySelector('[data-section="port"]'));
-            template.set('secure', itemTemplate.content.querySelector('[data-section="secure"]'));
-
-            /**
-             * adds connection item to list 
-             */
-            function addConnectionToList(data) {
-                for(section in data) {
-                    if (data.hasOwnProperty(section)) {
-                        template.get(section).textContent = data[section];
-                    }
-                }
-
-                const item = document.importNode(itemTemplate.content, true);
-                item.firstElementChild.addEventListener("click", (event) => {
-
-                    event.stopPropagation();
-                    event.preventDefault();
-
-                    const action = event.target.getAttribute("data-action");
-                    switch (action) {
-                        case "delete":
-                            vscode.postMessage({command: "delete", data: data});
-                            break;
-                    }
-                });
-
-                list.appendChild(item);
-            }
-
-            /**
-             * handle form add submit
-             */
-            function createConnection(event) {
-                event.preventDefault();
-                event.stopPropagation();
-
-                const connectionData = controls.reduce((data, control) => {
-                    const name  = control.getAttribute("data-control");
-                    const field = control.querySelector("input");
-                    const value = field.value;
-
-                    data[name] = 
-                        field.type === "checkbox" ? field.checked : 
-                        field.type === "number" ? parseInt(value, 10) : value;
-
-                    // reset values
-                    field.type === "checkbox" ? field.checked = true : field.value = "";
-                    return data;
-                }, {});
-                vscode.postMessage({ command: "add", data: connectionData });
-            };
-
-            /**
-             * handle commans from vscode
-             */
-            function handleVSCodeVCommand(event) {
-                const data = event.data;
-
-                switch(data.command) {
-                    case 'update':
-                        list.innerHTML = ""; // clear full content in list before not really good
-                        data.connections.forEach((connection) => addConnectionToList(connection))
-                        break;
-                }
-            }
-
-            /**
-             * bootstrap
-             */
-            function registerEvents() {
-                form.addEventListener("submit", createConnection);
-                window.addEventListener("message", handleVSCodeVCommand);
-            }
-
-            registerEvents();
-        </script>
-    </body>
-</html>
