+++ src/extension/data/tokens.ts
@@ -1,0 +1,10 @@
+import * as vscode from "vscode";
+
+export class SessionToken<T>  {
+    public constructor(private description: string = '') {}
+}
+
+export const ExtensionContext   = new SessionToken<vscode.ExtensionContext>("VSCode Extension Context");
+export const ConnectionSettings = new SessionToken<string>("VSCode Connection Settings");
+export const WorkspaceFolders   = new SessionToken<string[]>("VSCode Connection Settings");
+export const ExtensionPath      = new SessionToken<string>("VCode Extension Path");
