+++ src/extension/libs/qixfs/index.ts
@@ -1,0 +1,3 @@
+export * from "./src/qixfs.module";
+export * from "./src/entry";
+export * from "./src/utils";
