+++ src/extension/libs/qixfs/src/entry/index.ts
@@ -1,0 +1,8 @@
+export * from "./src/entry";
+export * from "./src/script/directory";
+export * from "./src/script/file";
+export * from "./src/variable/directory";
+export * from "./src/variable/file";
+export * from "./src/directory/app";
+export * from "./src/directory/documents";
+export * from "./src/file/temporary";
