+++ src/extension/libs/qixfs/src/entry/src/file/temporary.ts
