+++ src/extension/libs/qixfs/src/entry/src/script/directory.ts
