+++ src/extension/libs/qixfs/src/entry/src/script/file.ts
