+++ src/extension/libs/qixfs/src/entry/src/variable/directory.ts
