+++ src/extension/libs/qixfs/src/entry/src/variable/file.ts
