+++ src/extension/libs/qixfs/src/utils/src/workspace-folder-manager.ts
