+++ src/extension/libs/settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./src/settings.module";
