--- src/extension/qixfs/entry/index.ts
@@ -1,2 +1,0 @@
-export * from "./src/entry";
-export * from "./src/routes";
