--- src/extension/utils/src/bootstrap.ts
@@ -1,2 +1,0 @@
-export function bootstrap() {
-}
