+++ src/extension/utils/src/session-cache.ts
@@ -1,15 +1,6 @@
-import * as vscode from "vscode";
+import { SessionToken } from "@data/tokens";
 
 declare type R = any;
-
-export class SessionToken<T>  {
-    public constructor(private description: string = '') {}
-}
-
-export const ExtensionContext   = new SessionToken<vscode.ExtensionContext>("VSCode Extension Context");
-export const ConnectionSettings = new SessionToken<string>("VSCode Connection Settings");
-export const WorkspaceFolders   = new SessionToken<string[]>("VSCode Connection Settings");
-export const ExtensionPath      = new SessionToken<string>("VCode Extension Path");
 
 export class SessionCache {
